feat: cleaning parsers
