+++ ciphering-cli-tool.js
@@ -3,9 +3,8 @@ const errorHandler = require('./src/errors/errorHandler');
 const encryption = require('./src/ciphers/encryption');
 
 try {
-  validFlags(process.argv.slice(2));
+  const params = validFlags(process.argv.slice(2));
+  encryption(params);
 } catch(e) {
   errorHandler(e);
 }
-
-encryption(process.argv.slice(2));
