+++ output.txt
@@ -1,1 +1,1 @@
-Myxn xn nbdobm. Tbnnfzb ferlm "_" nhteru!Vhgw gw wkmxkv. Ckwwoik onauv "_" wqcnad!Hvwg wg gsqfsh. Asggous opcih "_" gmapcz!This is secret. Message about "_" symbol!
+Myxn xn nbdobm. Tbnnfzb ferlm "_" nhteru!
