+++ src/validFlags/validFlags.js
@@ -1,23 +1,26 @@
 const flags = require('../asssets/flags');
 const parsingFlags = require('./parsingFlags');
 const CustomError = require('../errors/customError');
-const validConfig = require('./validConfig');
+const parsingConfig = require('./parsingConfig');
 const validFile = require('./validFile');
 
 
 const validFlags = (argv) => {
   const parseFlags = parsingFlags(argv);
+  const result = { ...parseFlags };
 
   if (!parseFlags.config) throw new CustomError('The config is missing');
   Object.keys(parseFlags).forEach(key => {
     const value = parseFlags[key];
     if (key === 'config') {
-      validConfig(value);
+      result.config = parsingConfig(value);
     }
     if (key === 'input' || key === 'output') {
       validFile(value);
     }
   });
+
+  return result;
 };
 
 module.exports = validFlags;
